+++ app/back-end/util.js
@@ -3,15 +3,13 @@
 // 2016-18 Claus Zinn, University of Tuebingen
 // 
 // File: util.js
-// Time-stamp: <2018-06-28 12:24:11 (zinn)>
+// Time-stamp: <2018-06-29 18:40:49 (zinn)>
 // -------------------------------------------
 
 export const inclToolsReqAuth = process.env.INCL_TOOLS_REQ_AUTH;
-export const allowTextInput   = process.env.ALLOW_TEXT_INPUT;
-export const allowPasteURL    = process.env.ALLOW_PASTE_URL;
 export const fileStorage      = process.env.FILE_STORAGE;
-export const b2drop_user      = process.env.B2DROP_USER;
-export const b2drop_pass      = process.env.B2DROP_PASS;
+export const nextcloud_user   = process.env.NEXTCLOUD_USER;
+export const nextcloud_pass   = process.env.NEXTCLOUD_PASS;
 export const lrsVersion       = process.env.VERSION;
 export const emailContact     = process.env.CONTACT;
 export const appContextPath   = process.env.APP_CONTEXT_PATH;
@@ -20,37 +18,24 @@ export const emailContactCommand = "mailto:"+emailContact+"?subject=CLARIN-PLUS 
 // for creation of link
 export const fileStorageServerMPG_remote     = 'http://ws1-clarind.esc.rzg.mpg.de/drop-off/storage/';
 
-// see nginx.conf for reverse-proxying (all pathnames are local0
+// see nginx.conf for reverse-proxying (all pathnames are local)
 export const fileStorageServerMPG_localhost       = '/storage/';
 export const fileStorageServerNEXTCLOUD_localhost = '/nextcloud/';
-export const fileStorageServerB2DROP_localhost    = '/btwodrop/';
 
 export function unfoldHandle( handle ) {
     var hdlShortPrefix = "hdl:";
     var protocol = window.location.protocol;
     var hdlLongPrefix  = protocol.concat("//hdl.handle.net/");	
     var index = handle.indexOf(hdlShortPrefix);
-
     var expandedHandle = decodeURIComponent(handle);
-
-    console.log('util/unfoldHandle', expandedHandle)
 
     if (index > -1) {
 	expandedHandle = hdlLongPrefix.concat( handle.substring(index+hdlShortPrefix.length, handle.length) );
-    } else {
-	console.log('util.js/unfoldHandle: there is no need to unfold the handle', handle);
+	console.log('util/unfoldHandle expanded handle', handle, expandedHandle)
     }
     return expandedHandle;
 }
 
-/* This is a CORS relict, see the rewriting in nginx.conf.
-   - Note that B2DROP does not work with this 'trick', so here's a (browser-external)
-     Python script is used for downloading 
-   - iCloud does not provide a direct link to access a shared file (login etc. required)
-   - b2drop.eudat.eu is not active yet, but there is a test instance at https://fsd-cloud48.zam.kfa-juelich.de
-   - the first if condition shadows all subsequent nextcloud-based conditions
-   - paste events are always fetched via the Python script
-*/
 
 export function fileExtensionChooser (mimetype) {
     var extension = "txt";
@@ -79,14 +64,24 @@ export function fileExtensionChooser (mimetype) {
     return extension;
 }
 
-export function rewriteURL( caller, fileURL ) {
-    var href = window.location.origin.concat(window.location.pathname);
+/* This is a CORS relict, see the rewriting in nginx.conf.
+   - Note that B2DROP does not work with this 'trick', so here's a (browser-external)
+     Python script is used for downloading 
+   - iCloud does not provide a direct link to access a shared file (login etc. required)
+   - b2drop.eudat.eu is not active yet, but there is a test instance at https://fsd-cloud48.zam.kfa-juelich.de
+   - the first if condition shadows all subsequent nextcloud-based conditions
+   - paste events are always fetched via the Python script
+*/
+
+export function rewriteURL( fileURL ) {
     var windowAppContextPath = window.APP_CONTEXT_PATH;
 
-    console.log('util/rewriteURL at start', caller, fileURL, href, windowAppContextPath);
+    // todo: delete rest of body
+    return windowAppContextPath.concat('/download?input='+encodeURI(fileURL));
 
+    var href = window.location.origin.concat(window.location.pathname);
     var corsLink = "";
-    if (caller == "B2DROP") {
+    if ( (caller == "B2DROP") || (caller == "VLO"))  {
 	// nop   -- fileURL = fileURL.concat('/download')
     } else if ( caller == "PASTE") {
 	// nop
@@ -107,15 +102,13 @@ export function rewriteURL( caller, fileURL ) {
 	corsLink = fileURL;
     }
 
-    console.log('util/rewriteURL at end', caller, fileURL, href, windowAppContextPath);
+    console.log('util/rewriteURL', caller, fileURL, href, windowAppContextPath, corsLink);
 
     // todo: harmonize this with the use of APP_CONTEXT_PATH in main App component.
-    if ( (caller == "B2DROP") || (fileURL.indexOf('hdl.handle.net') > 1) || (caller == "PASTE") ) {
+    if ( (caller == "B2DROP") || (fileURL.indexOf('hdl.handle.net') > 1) || (caller == "PASTE") || true ) {
 	return windowAppContextPath.concat('/download?input='+encodeURI(fileURL));
-	//return href.concat('download?input='+encodeURI(fileURL)) // the reverse proxy to the python script
     } else {
 	return windowAppContextPath.concat('/').concat(corsLink);
-//	return href.concat(corsLink);                            // the reverse proxy to the clouds
     }
 }