+++ app/components/Resources.jsx
@@ -1,3 +1,11 @@
+// -------------------------------------------
+// The CLARIN Language Resource Switchboard
+// 2016-18 Claus Zinn, University of Tuebingen
+// 
+// File: Resources.jsx
+// Time-stamp: <2018-06-29 20:23:59 (zinn)>
+// -------------------------------------------
+
 import React from 'react';
 import Resource from './Resource.jsx';