+++ app/components/Tool.jsx
@@ -1,3 +1,11 @@
+// -------------------------------------------
+// The CLARIN Language Resource Switchboard
+// 2016-18 Claus Zinn, University of Tuebingen
+// 
+// File: Tool.jsx
+// Time-stamp: <2018-06-29 20:24:25 (zinn)>
+// -------------------------------------------
+
 import React from 'react';
 // import ReactTooltip from 'react-tooltip';
 import Accordion from '../helperComponents/Accordion';